python/tests/requirements_tests.txt
pytest
networkx>=2.6.3
requests
pyjwt>=2.2.0
rich>=12.0.0
graphviz>=0.20.1